pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cisa-for-bdsca"
version = "0.1.0"
description = "Collect CISA vulnerability information from Black Duck SCA for CVE, EUVD, and BDSA vulnerabilities"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jouni Lehto", email = "lehto.jouni@gmail.com"}
]
keywords = ["security", "vulnerabilities", "cisa", "black-duck", "sca", "cve", "euvd", "bdsa"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "blackduck>=1.1.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "typer[all]>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
]

[project.scripts]
cisa-bdsca = "cisa_bdsca.cli:app"

[project.urls]
Homepage = "https://github.com/lejouni/cisa-for-bdsca"
Repository = "https://github.com/lejouni/cisa-for-bdsca"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/cisa_bdsca --cov-report=term-missing --cov-report=html"
